requirements.txt
pipettor>=0.4.0
numpy
biopython
enum-compat
mysqlclient
apsw
deprecation
flake8
pyBigWig
